chore(demo): change order of example in documentation page for `InputNumber`
