+++ projects/demo/src/modules/components/input-number/examples/10/index.ts
@@ -1,29 +1,16 @@
-import {Component, computed, signal} from '@angular/core';
+import {JsonPipe} from '@angular/common';
+import {Component} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import {tuiNumberFormatProvider} from '@taiga-ui/core';
-import {TuiInputNumber, tuiInputNumberOptionsProvider} from '@taiga-ui/kit';
+import {TuiInputNumber} from '@taiga-ui/kit';
 
 @Component({
-    imports: [FormsModule, TuiInputNumber],
+    imports: [FormsModule, JsonPipe, TuiInputNumber],
     templateUrl: './index.html',
     encapsulation,
     changeDetection,
-    providers: [
-        tuiInputNumberOptionsProvider({
-            min: -Infinity,
-            max: Infinity,
-        }),
-        tuiNumberFormatProvider({precision: 0}),
-    ],
 })
 export default class Example {
-    protected readonly value = signal<bigint | null>(
-        BigInt(`777${Number.MAX_SAFE_INTEGER}00`),
-    );
-
-    protected readonly stringified = computed((x = this.value()) =>
-        typeof x === 'bigint' ? `${x}n` : 'null',
-    );
+    protected value = 0.5;
 }
