+++ projects/demo/src/modules/components/input-number/examples/2/index.html
@@ -1,19 +1,14 @@
-<tui-textfield
-    iconStart="@tui.euro"
-    tuiTextfieldSize="m"
->
-    <label tuiLabel>I am a label</label>
+<tui-textfield>
+    <label tuiLabel>Enter a really huge integer</label>
 
     <input
-        placeholder="I am placeholder"
+        bigint
         tuiInputNumber
         [(ngModel)]="value"
     />
-
-    <tui-icon
-        icon="@tui.circle-alert"
-        style="color: var(--tui-status-negative)"
-    />
-
-    <tui-icon tuiTooltip="I am a hint" />
 </tui-textfield>
+
+<p>
+    <strong>Control value:</strong>
+    <code>{{ stringified() }}</code>
+</p>
