+++ projects/demo/src/modules/components/input-number/examples/4/index.html
@@ -1,10 +1,14 @@
 <tui-textfield>
-    <label tuiLabel>Price</label>
+    <label tuiLabel>Type number like a German</label>
 
     <input
         tuiInputNumber
-        [min]="0"
-        [prefix]="'USD' | tuiCurrency"
+        [tuiNumberFormat]="numberFormat"
         [(ngModel)]="value"
+    />
+
+    <tui-icon
+        tuiHintDirection="right"
+        tuiTooltip="In Germany people use comma as decimal separator and point for thousands"
     />
 </tui-textfield>
