+++ projects/demo/src/modules/components/input-number/examples/4/index.ts
@@ -2,15 +2,19 @@ import {Component} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import {TuiCurrencyPipe} from '@taiga-ui/addon-commerce';
-import {TuiInputNumber} from '@taiga-ui/kit';
+import {TuiIcon, TuiNumberFormat, type TuiNumberFormatSettings} from '@taiga-ui/core';
+import {TuiInputNumber, TuiTooltip} from '@taiga-ui/kit';
 
 @Component({
-    imports: [FormsModule, TuiCurrencyPipe, TuiInputNumber],
+    imports: [FormsModule, TuiIcon, TuiInputNumber, TuiNumberFormat, TuiTooltip],
     templateUrl: './index.html',
     encapsulation,
     changeDetection,
 })
 export default class Example {
-    protected value: number | null = null;
+    protected value: number | null = 1_234_567.89;
+    protected numberFormat: Partial<TuiNumberFormatSettings> = {
+        decimalSeparator: ',',
+        thousandSeparator: '.',
+    };
 }
