+++ projects/demo/src/modules/components/input-number/examples/5/index.html
@@ -1,12 +1,10 @@
 <tui-textfield>
-    <label tuiLabel>Percentage</label>
+    <label tuiLabel>Price</label>
 
     <input
-        postfix="%"
         tuiInputNumber
-        [max]="100"
         [min]="0"
-        [step]="1"
+        [prefix]="'USD' | tuiCurrency"
         [(ngModel)]="value"
     />
 </tui-textfield>
