+++ projects/demo/src/modules/components/input-number/examples/5/index.ts
@@ -2,10 +2,11 @@ import {Component} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
+import {TuiCurrencyPipe} from '@taiga-ui/addon-commerce';
 import {TuiInputNumber} from '@taiga-ui/kit';
 
 @Component({
-    imports: [FormsModule, TuiInputNumber],
+    imports: [FormsModule, TuiCurrencyPipe, TuiInputNumber],
     templateUrl: './index.html',
     encapsulation,
     changeDetection,
