+++ projects/demo/src/modules/components/input-number/examples/7/index.html
@@ -1,9 +1,29 @@
-<tui-textfield [style.width.rem]="10">
+<tui-textfield [tuiTextfieldCleaner]="false">
     <input
-        postfix=" €"
-        tuiFluidTypography
         tuiInputNumber
         [min]="0"
         [(ngModel)]="value"
+        (keydown.arrowDown)="onStep(-1)"
+        (keydown.arrowUp)="onStep(+1)"
     />
+
+    <button
+        appearance="floating"
+        size="s"
+        tuiButton
+        type="button"
+        (click.prevent)="onStep(+100)"
+    >
+        + 100
+    </button>
+
+    <button
+        appearance="floating"
+        size="s"
+        tuiButton
+        type="button"
+        (click.prevent)="onStep(+1000)"
+    >
+        + 1000
+    </button>
 </tui-textfield>
