+++ projects/demo/src/modules/components/input-number/examples/7/index.less
