+++ projects/demo/src/modules/components/input-number/examples/8/index.ts
@@ -2,30 +2,14 @@ import {Component} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import {TuiValueTransformer} from '@taiga-ui/cdk';
-import {TuiInputNumber, tuiInputNumberOptionsProvider} from '@taiga-ui/kit';
-
-class NaNTransformer extends TuiValueTransformer<number | null, number> {
-    public override fromControlValue(value: number): number | null {
-        return Number.isNaN(value) ? null : value;
-    }
-
-    public override toControlValue(value: number | null): number {
-        return value ?? NaN;
-    }
-}
+import {TuiFluidTypography, TuiInputNumber} from '@taiga-ui/kit';
 
 @Component({
-    imports: [FormsModule, TuiInputNumber],
+    imports: [FormsModule, TuiFluidTypography, TuiInputNumber],
     templateUrl: './index.html',
     encapsulation,
     changeDetection,
-    providers: [
-        tuiInputNumberOptionsProvider({
-            valueTransformer: new NaNTransformer(),
-        }),
-    ],
 })
 export default class Example {
-    protected value = NaN;
+    protected value = 10000;
 }
