+++ projects/demo/src/modules/components/input-number/examples/9/index.html
@@ -1,12 +1,12 @@
 <tui-textfield>
     <input
+        placeholder="Form control contains NaN"
         tuiInputNumber
-        [max]="1"
-        [min]="0"
-        [quantum]="0.05"
         [(ngModel)]="value"
     />
 </tui-textfield>
 
-<p><strong>Control value:</strong></p>
-<code>{{ value | json }}</code>
+<p>
+    Control value:
+    <code>{{ value.toString() }}</code>
+</p>
