+++ projects/demo/src/modules/components/input-number/index.html
@@ -24,10 +24,36 @@
         </tui-doc-example>
 
         <tui-doc-example
-            id="texfield-based"
-            heading="Textfield-based"
+            id="bigint-type-control-value"
+            heading="BigInt as form control value"
             [component]="2 | tuiComponent"
             [content]="2 | tuiExample"
+            [description]="bigintFormControlDescription"
+        >
+            <ng-template #bigintFormControlDescription>
+                Form control value type can be changed to
+                <code>bigint</code>
+                .
+
+                <p>
+                    Just put
+                    <code>bigint</code>
+                    attribute and set new values for
+                    <code>min</code>
+                    ,
+                    <code>max</code>
+                    and
+                    <code>precision</code>
+                    properties.
+                </p>
+            </ng-template>
+        </tui-doc-example>
+
+        <tui-doc-example
+            id="texfield-based"
+            heading="Textfield-based"
+            [component]="3 | tuiComponent"
+            [content]="3 | tuiExample"
             [description]="textfieldBasedDescription"
         >
             <ng-template #textfieldBasedDescription>
@@ -57,8 +83,8 @@
         <tui-doc-example
             id="localization"
             heading="Localization"
-            [component]="3 | tuiComponent"
-            [content]="3 | tuiExample"
+            [component]="4 | tuiComponent"
+            [content]="4 | tuiExample"
             [description]="numberFormatDescription"
         >
             <ng-template #numberFormatDescription>
@@ -76,8 +102,8 @@
         <tui-doc-example
             id="affixes"
             heading="Affixes"
-            [component]="4 | tuiComponent"
-            [content]="4 | tuiExample"
+            [component]="5 | tuiComponent"
+            [content]="5 | tuiExample"
             [description]="affixesDescription"
         >
             <ng-template #affixesDescription>
@@ -102,8 +128,8 @@
         <tui-doc-example
             id="step"
             heading="Step"
-            [component]="5 | tuiComponent"
-            [content]="5 | tuiExample"
+            [component]="6 | tuiComponent"
+            [content]="6 | tuiExample"
             [description]="stepDescription"
         >
             <ng-template #stepDescription>
@@ -121,15 +147,15 @@
         <tui-doc-example
             id="custom-step-buttons"
             heading="Custom step buttons"
-            [component]="6 | tuiComponent"
-            [content]="6 | tuiExample"
+            [component]="7 | tuiComponent"
+            [content]="7 | tuiExample"
         />
 
         <tui-doc-example
             id="fluid-typography"
             heading="Fluid typography"
-            [component]="7 | tuiComponent"
-            [content]="7 | tuiExample"
+            [component]="8 | tuiComponent"
+            [content]="8 | tuiExample"
             [description]="fluidTypographyDescription"
         >
             <ng-template #fluidTypographyDescription>
@@ -148,8 +174,8 @@
         <tui-doc-example
             id="value-transformer"
             heading="Value transformer"
-            [component]="8 | tuiComponent"
-            [content]="8 | tuiExample"
+            [component]="9 | tuiComponent"
+            [content]="9 | tuiExample"
             [description]="transformerDescription"
         >
             <ng-template #transformerDescription>
@@ -170,8 +196,8 @@
         <tui-doc-example
             id="quantum"
             heading="Quantum"
-            [component]="9 | tuiComponent"
-            [content]="9 | tuiExample"
+            [component]="10 | tuiComponent"
+            [content]="10 | tuiExample"
             [description]="quantumDescription"
         >
             <ng-template #quantumDescription>
@@ -199,34 +225,6 @@
                     ,
                     <code>1</code>
                     .
-                </p>
-            </ng-template>
-        </tui-doc-example>
-
-        <!-- TODO: change order of examples in separate PR -->
-        <!-- (otherwise, snapshots of every previous examples will fail – easy to miss real bug) -->
-        <tui-doc-example
-            id="bigint-type-control-value"
-            heading="BigInt as form control value"
-            [component]="10 | tuiComponent"
-            [content]="10 | tuiExample"
-            [description]="bigintFormControlDescription"
-        >
-            <ng-template #bigintFormControlDescription>
-                Form control value type can be changed to
-                <code>bigint</code>
-                .
-
-                <p>
-                    Just put
-                    <code>bigint</code>
-                    attribute and set new values for
-                    <code>min</code>
-                    ,
-                    <code>max</code>
-                    and
-                    <code>precision</code>
-                    properties.
                 </p>
             </ng-template>
         </tui-doc-example>
